bee-163 Remove old blueprint
