+++ frontend-beeyond/app/src/app/modules/blueprint/blueprint-routing.module.ts
@@ -1,13 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { BlueprintComponent } from './pages/blueprint/blueprint.component';
-import { BlueprintTemplateComponent } from './pages/blueprint-template/blueprint-template.component';
-import { BlueprintNewComponent } from './pages/blueprint-new/blueprint-new.component';
 
 const routes: Routes = [
-  { path: '', component: BlueprintComponent },
-  { path: 'template/:id', component: BlueprintTemplateComponent },
-  { path: 'new', component: BlueprintNewComponent }
+  { path: '', component: BlueprintComponent }
 ];
 
 @NgModule({
