--- frontend-beeyond/app/src/app/modules/blueprint/pages/blueprint-template/blueprint-template.component.html
@@ -1,47 +1,0 @@
-<ng-container *ngIf="template">
-  <mat-card-title>{{ template.name }}</mat-card-title>
-  <mat-card-subtitle>{{ template.description }}</mat-card-subtitle>
-
-  <form [formGroup]="templateForm">
-    <div
-      *ngFor="let field of template.fields; let i = index"
-      class="field-container"
-      formArrayName="fieldValues"
-    >
-      <ng-container [formGroupName]="i">
-        <mat-form-field class="full-width">
-          <mat-label>{{ field.label }}</mat-label>
-          <input matInput formControlName="value" />
-          <mat-hint>{{ field.description }}</mat-hint>
-          <mat-error *ngIf="templateForm.get('fieldValues').invalid">
-            {{ field.description }}
-          </mat-error>
-        </mat-form-field>
-      </ng-container>
-    </div>
-
-    <mat-form-field class="field-container full-width">
-      <mat-label>An optional note:</mat-label>
-      <textarea
-        formControlName="note"
-        matInput
-        cdkTextareaAutosize
-        cdkAutosizeMinRows="1"
-        cdkAutosizeMaxRows="5"
-      ></textarea>
-      <mat-error *ngIf="templateForm.get('note').invalid">
-        The note must not be longer than 255 characters
-      </mat-error>
-    </mat-form-field>
-  </form>
-  <mat-card-actions align="right">
-    <button
-      mat-flat-button
-      color="primary"
-      (click)="submitApplication()"
-      [disabled]="!templateForm.valid"
-    >
-      Submit
-    </button>
-  </mat-card-actions>
-</ng-container>
