--- frontend-beeyond/app/src/app/modules/blueprint/pages/blueprint-template/blueprint-template.component.scss
@@ -1,7 +1,0 @@
-.field-container {
-  padding-top: 0.8rem;
-}
-
-::ng-deep .new-line {
-  white-space: pre-wrap;
-}
